+++ backend/src/Controller/Api/ExpensesListsController.php
@@ -15,6 +15,7 @@ final class ExpensesListsController extends AbstractController
   public function lists_get(ExpensesListsService $expensesListsService): JsonResponse
   {
     $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $lists = $expensesListsService->getLists($currentUser);
     if (!$lists) return $this->json(['error' => 'Lists not found'], 404);
     return $this->json($lists, 200);
@@ -24,6 +25,7 @@ final class ExpensesListsController extends AbstractController
   public function list_get(ExpensesListsService $expensesListsService, $id): JsonResponse
   {
     $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $list = $expensesListsService->getList($id, $currentUser);
     if (!$list) return $this->json(['error' => 'List not found'], 404);
     return $this->json($list, 200);
@@ -32,6 +34,8 @@ final class ExpensesListsController extends AbstractController
   #[Route('/', name: 'list_create', methods: ['POST'])]
   public function list_create(Request $request, ExpensesListsService $expensesListsService): JsonResponse
   {
+    $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $data = json_decode($request->getContent(), true);
     $list = $expensesListsService->addList($data);
     if (!$list) return $this->json(['error' => 'List already exists'], 403);
@@ -42,6 +46,7 @@ final class ExpensesListsController extends AbstractController
   public function list_update(Request $request, ExpensesListsService $expensesListsService, $id): JsonResponse
   {
     $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $data = json_decode($request->getContent(), true);
     $list = $expensesListsService->setList($data, $id, $currentUser);
     if (!$list) return $this->json(['error' => 'List not found']);
@@ -52,6 +57,7 @@ final class ExpensesListsController extends AbstractController
   public function list_delete(ExpensesListsService $expensesListsService, $id): JsonResponse
   {
     $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $deleted = $expensesListsService->deleteList($id, $currentUser);
     if (!$deleted) return $this->json(['error' => 'List not found'], 404);
