+++ backend/src/Controller/Api/KmMileageRatesController.php
@@ -14,6 +14,8 @@ final class KmMileageRatesController extends AbstractController
   #[Route('/', name: 'rates_get', methods: ['GET'])]
   public function rates_get(KmMileageRatesService $kmMileageRatesService): JsonResponse
   {
+    $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $rates = $kmMileageRatesService->getRates();
     if (!$rates) return $this->json(['error' => 'Rates not found'], 404);
     return $this->json($rates, 200);
@@ -22,6 +24,8 @@ final class KmMileageRatesController extends AbstractController
   #[Route('/{id}', name: 'rate_get', requirements: ['id' => '\d+'], methods: ['GET'])]
   public function rate_get(KmMileageRatesService $kmMileageRatesService, $id): JsonResponse
   {
+    $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $rate = $kmMileageRatesService->getRate($id);
     if (!$rate) return $this->json(['error' => 'Rate not found'], 404);
     return $this->json($rate, 200);
@@ -31,6 +35,7 @@ final class KmMileageRatesController extends AbstractController
   public function rate_create(Request $request, KmMileageRatesService $kmMileageRatesService): JsonResponse
   {
     $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $data = json_decode($request->getContent(), true);
     $rate = $kmMileageRatesService->addRate($data, $currentUser);
     if (!$rate) return $this->json(['error' => 'Rate already exists'], 409);
@@ -42,6 +47,7 @@ final class KmMileageRatesController extends AbstractController
   public function rate_update(Request $request, KmMileageRatesService $kmMileageRatesService, $id): JsonResponse
   {
     $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $data = json_decode($request->getContent(), true);
     $rate = $kmMileageRatesService->setRate($data, $id, $currentUser);
     if (!$rate) return $this->json(['error' => 'Rate not found']);
@@ -53,6 +59,7 @@ final class KmMileageRatesController extends AbstractController
   public function document_delete(KmMileageRatesService $kmMileageRatesService, $id): JsonResponse
   {
     $currentUser = $this->getUser();
+    if (!$currentUser) return $this->json(['error' => 'Not connected'], 401);
     $deleted = $kmMileageRatesService->deleteRate($id, $currentUser);
     if (!$deleted) return $this->json(['error' => 'Rate not found'], 404);
     if ($deleted === 'Forbidden') return $this->json(['error' => 'Delete forbidden'], 403);
